Move the Ollama API handlers into serve/ollama.rs
